pyproject.toml
[project]
name = "aitravelplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amadeus>=12.0.0",
    "arxiv>=2.2.0",
    "bs4>=0.0.2",
    "defusedxml>=0.7.1",
    "duckduckgo-search>=8.0.4",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "geopy>=2.4.1",
    "google-maps-places>=0.2.1",
    "google-search-results>=2.4.2",
    "googlemaps>=4.10.0",
    "hf-xet>=1.1.4",
    "htmltabletomd>=1.0.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-experimental>=0.3.4",
    "langchain-gemini>=0.1.1",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-milvus>=0.2.0",
    "langchain-openai>=0.3.24",
    "langchain-pinecone>=0.2.8",
    "langgraph>=0.4.8",
    "libmagic>=1.0",
    "numpy>=2.3.0",
    "open-clip-torch>=2.32.0",
    "pandas>=2.3.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pi-heif>=0.22.0",
    "pinecone>=7.2.0",
    "poppler-utils>=0.1.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "sentence-transformers>=4.1.0",
    "serpapi>=0.1.5",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
    "unstructured-inference>=1.0.5",
    "unstructured[all-docs]>=0.17.2",
    "wikipedia>=1.4.0",
    "xmltodict>=0.14.2",
    "yfinance>=0.2.63",
    "youtube-search>=2.1.2",
]
